requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
gast==0.2.0
GPUtil==1.3.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
Werkzeug==0.14.1
